Cleanup from Review
